Refactored code in service layer

+++ src/main/java/com/everest/employeeportal/controller/EmployeeController.java
@@ -1,5 +1,6 @@
 package com.everest.employeeportal.controller;
 
+import com.everest.employeeportal.exceptions.EmployeeNotFoundException;
 import com.everest.employeeportal.exceptions.RequiredRequestParamException;
 import com.everest.employeeportal.models.ApiResponse;
 import com.everest.employeeportal.models.Employee;
@@ -30,7 +31,10 @@ public class EmployeeController{
     }
     @GetMapping("/{empId}")
     public Employee getEmployeeById(@PathVariable("empId") Long empId){
-            return employeeService.getEmployeeById(empId);
+           if(employeeService.getEmployeeById(empId).isPresent()){
+               return employeeService.getEmployeeById(empId).get();
+           }
+           throw new EmployeeNotFoundException(empId);
     }
 
     @PostMapping("")
@@ -47,6 +51,10 @@ public class EmployeeController{
 
     @DeleteMapping("/{empId}")
     public ResponseEntity<Object> deleteEmployee(@PathVariable("empId") Long empId) {
+        if(employeeService.getEmployeeById(empId).isPresent()){
+            ApiResponse apiResponse = new ApiResponse("employee not found with Id "+empId);
+            return new ResponseEntity<>(apiResponse, HttpStatus.NOT_FOUND);
+        }
         employeeService.deleteEmployee(empId);
         ApiResponse apiResponse = new ApiResponse("Deleted employee with Id  "+empId);
         return new ResponseEntity<>(apiResponse, HttpStatus.OK);
